third_party/WebKit/LayoutTests/svg/css/parse-length.html

<!DOCTYPE html>
<html>
<style>
* { font-size: 16px; }
svg, rect { font-family: 'Ahem'; }
div { font-size: 8px; } 
</style>
<html>
<svg id="svg" width="0" height="0"></svg>
<script src="../../resources/js-test.js"></script>
<script>
description("Test that 'length' presentation attribute values are parsed with CSS presentation rules.");

function computedStyle(property, value) {
    var rect = document.createElementNS('http://www.w3.org/2000/svg', 'rect');
    document.getElementById('svg').appendChild(rect);
    rect.setAttribute(property, value);
    var computedValue = getComputedStyle(rect).getPropertyValue(property);
    document.getElementById('svg').removeChild(rect);
    return computedValue;
}

function testComputed(property, value, expected) {
    shouldBeEqualToString('computedStyle("' + property + '", "' + value + '")', expected);
}

function negativeTest(property, value) {
    testComputed(property, value, "auto");
}

function negativeTestZero(property, value) {
    testComputed(property, value, "0px");
}

// Test 'x'.
testComputed("x", "  100", "100px");
testComputed("x", "100   ", "100px");
testComputed("x", "100px", "100px");
testComputed("x", "1em", "16px");
// testComputed("x", "1ex", "12.8000001907349px"); // enable this again once http://crbug.com/441840 is fixed
testComputed("x", "20%", "20%");
testComputed("x", "-200px", "-200px");

// Test 'y'.
testComputed("y", "  100", "100px");
testComputed("y", "100   ", "100px");
testComputed("y", "100px", "100px");
testComputed("y", "1em", "16px");
// testComputed("y", "1ex", "12.8000001907349px"); // enable this again once http://crbug.com/441840 is fixed
testComputed("y", "20%", "20%");
testComputed("y", "-200px", "-200px");

// Negative tests for 'x'.
negativeTestZero("x", "auto", "auto");
negativeTestZero("x", "100   px");
negativeTestZero("x", "100px;");
negativeTestZero("x", "100px !important");
negativeTestZero("x", "{ 100px }");

// Negative tests for 'y'.
negativeTestZero("y", "auto");
negativeTestZero("y", "100   px");
negativeTestZero("y", "100px;");
negativeTestZero("y", "100px !important");
negativeTestZero("y", "{ 100px }");

// Test 'rx'.
testComputed("rx", "  100", "100px");
testComputed("rx", "100   ", "100px");
testComputed("rx", "100px", "100px");
testComputed("rx", "1em", "16px");
// testComputed("rx", "1ex", "12.8000001907349px"); // enable this again once http://crbug.com/441840 is fixed
testComputed("rx", "20%", "20%");
testComputed("rx", "-200px", "-200px");

// Test 'ry'.
testComputed("ry", "  100", "100px");
testComputed("ry", "100   ", "100px");
testComputed("ry", "100px", "100px");
testComputed("ry", "1em", "16px");
// testComputed("ry", "1ex", "12.8000001907349px"); // enable this again once http://crbug.com/441840 is fixed
testComputed("ry", "20%", "20%");
testComputed("ry", "-200px", "-200px");

// Negative tests for 'rx'.
negativeTestZero("rx", "auto", "auto");
negativeTestZero("rx", "100   px");
negativeTestZero("rx", "100px;");
negativeTestZero("rx", "100px !important");
negativeTestZero("rx", "{ 100px }");

// Negative tests for 'ry'.
negativeTestZero("ry", "auto");
negativeTestZero("ry", "100   px");
negativeTestZero("ry", "100px;");
negativeTestZero("ry", "100px !important");
negativeTestZero("ry", "{ 100px }");
</script>
</body>
</html>
